refactor: moved retry into root of utils module
